projects(tests): fix project sync tests
